requirements.txt
Flask==2.3.3
tensorflow==2.12.0
opencv-python==4.9.0.74
Pillow==10.1.0
numpy==1.25.0
reportlab==4.0.0
Werkzeug==2.3.7
